agent-hotel/requirements.txt
livekit-agents[openai,silero,groq]==1.4.1
livekit-plugins-turn-detector==1.4.1
python-dotenv==1.0.1
aiohttp==3.12.15
numpy
